requirements.txt
Flask[async]==2.3.2
requests==2.31.0
aiohttp==3.8.5
googleapis-common-protos==1.61.0
pycryptodome==3.19.0
protobuf==4.25.1
Werkzeug==2.3.7
gunicorn==20.1.0